requirements.txt
Flask
psutil
scapy
speedtest-cli
requests
plotly

